pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
torchsofa = ["conf/*.yaml"]

[project]
name = "torchsofa"
version = "2.0.0"
dependencies = [
    "hydra-core",
    # "torch",
    # "numpy",
    # "pandas",
    # "scipy",
    # "tqdm",
    # "matplotlib",
    # "pytorch-lightning",
]

[project.scripts]
sofa-infer = "torchsofa.cli:infer"
sofa-train = "torchsofa.cli:train"
sofa-eval = "torchsofa.cli:eval"
